requirements.txt
python-telegram-bot==10.0.0
tabulate==0.8.2
